feat/RSS-ECOMM-4_08: replace add to cart functionality to separate component, implement remove from cart
