+++ src/pages/Product/Product.page.module.scss
@@ -64,20 +64,14 @@
   width: 230px;
   display: flex;
   flex-direction: row;
-  justify-content: center;
-  gap: 24px;
+  justify-content: space-between;
+
   margin-top: 25px;
   padding: 8px;
   box-shadow:
     rgba(0, 0, 0, 0.2) 0px 8px 10px -5px,
     rgba(0, 0, 0, 0.14) 0px 16px 24px 2px,
     rgba(0, 0, 0, 0.12) 0px 6px 30px 5px;
-}
-
-.add-btn svg,
-.remove-btn svg {
-  width: 30px;
-  height: 30px;
 }
 
 @media (max-width: 950px) {
